Fix duplicate alias error when merging multiple scopes
When we try to merge multiple scopes from the same table, there's a
duplicate alias error because the alias is generated from the table name
only.

The increment mechanism currently in place doesn't work in this case as
we are merging scopes and AR has no mechanism in place to make these
relation objects communicate between them.

This patch simply pass around the scope_name down to the
`pg_search_rank_table_alias` method so it can be used as a component for
generating the alias name. There's a drawback though, the
`with_pg_search_rank` must be passed the scope_name argument to pass it
down, thus the API is not compatible.